requirements.txt
crewai==1.7.2
faiss-cpu==1.13.2
google-generativeai==0.8.6
langchain-community==0.4.1
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
pypdf==6.9.2
python-dotenv==1.1.1
streamlit==1.51.0
